requirements-core.txt
python-dotenv>=1.0
typer>=0.12
rich>=13.7
pytest>=8.3
ruff>=0.6
